[PloobsEngine] New Content Loader supporting Async loading and Unload of a single asset [PloobsEngine] CleanUp Pattern implemented [PloobsEngine] Id is now a Long [PloobsEngine] Entity Mapper now uses Weak References [Ploobs WP7] Fixed a bug when changing orientation

+++ IntroductionDemo/IntroductionDemo4.0/IntroductionDemo4.0/IntroductionDemo4.0/Demo6 - Trigger/TriggerBepuScreen.cs
@@ -258,13 +258,13 @@ namespace IntroductionDemo4._0
         #endregion
 
         #region IEntity Members
-        int id;
-        public int GetId()
+        long id;
+        public long GetId()
         {
             return id;
         }
 
-        public void SetId(int id)
+        public void SetId(long id)
         {
             this.id = id;
         }
